Adds Marowak, fixes unicode issue
Previously, unicode strings were not supported, so Pokémon had to be Pokemon, the issue seems to have been caused by the print method not getting the information that it was a unicode string from the __str__ method. The method has now been renamed to get_Pokedex_Message, and resolves the issue.