Fixes issue #1
Corrects the issue with Missingno, which was caused by using PIL to read the image into python in the converter. This is achieved by using scipy and the misc.imread function instead, with the mode set to "RGB". This caused the image x,y to be inversed from the PIL object.